+++ src/plan/lxr/rc.rs
@@ -306,7 +306,7 @@ impl<VM: VMBinding, const KIND: EdgeKind> ProcessIncs<VM, KIND> {
                         //Originaly was : let _ = self.rc.inc(target);
                         let result = self.rc.inc(target);
                         debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(target.to_raw_address(), Ordering::SeqCst) != 0);
-                        debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(target.to_raw_address(), Ordering::SeqCst) as u16 <= self.rc.count(target));
+                        debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(target.to_raw_address(), Ordering::SeqCst) <= self.rc.count(target));
                         #[cfg(feature = "measure_rc_rate")]
                         {
                             self.inc_objs += 1;
@@ -512,7 +512,7 @@ impl<VM: VMBinding, const KIND: EdgeKind> ProcessIncs<VM, KIND> {
         }
 
         debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(new.to_raw_address(), Ordering::SeqCst) != 0);
-        debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(new.to_raw_address(), Ordering::SeqCst) as u16 <= self.rc.count(new));
+        debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(new.to_raw_address(), Ordering::SeqCst) <= self.rc.count(new));
         // Put this into remset if this is a mature slot, or a weak root
         if K != EDGE_KIND_ROOT || add_root_to_remset {
             self.record_mature_evac_remset(s, new);
@@ -1107,7 +1107,7 @@ impl<VM: VMBinding> ProcessDecs<VM> {
                     }
                 }
                 debug_assert!(s_rc_prev_val != 0 || s_candidates.contains(&o));
-                debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(o.to_raw_address(), Ordering::SeqCst) as u16 <= self.rc.count(o) 
+                debug_assert!(STRONG_RC_TABLE.load_atomic::<u8>(o.to_raw_address(), Ordering::SeqCst) <= self.rc.count(o) 
                             || s_candidates.contains(&o));
             }
             if crate::args::PREFETCH {
